+++ src/StatisticsFunctions/kernels/plp_var_q16s_xpulpv2.c
@@ -73,19 +73,18 @@ void plp_var_q16s_xpulpv2(
                          const int16_t * __restrict__ pSrc,
                          uint32_t blockSize,
                          uint32_t deciPoint,
-                         int32_t * __restrict__ pRes){
+                         int16_t * __restrict__ pRes){
     
- 
   int32_t square_of_mean;
   int32_t square_of_values;
 
   int16_t mean;
 
   plp_mean_i16(pSrc, blockSize, &mean);
-  square_of_mean = (mean * mean) >> deciPoint;
+  square_of_mean = ((int32_t)mean * (int32_t)mean) >> deciPoint;
 
   plp_power_q16(pSrc, blockSize, deciPoint, &square_of_values);
   
-  *pRes = (square_of_values/blockSize - square_of_mean);
+  *pRes = (int16_t)(square_of_values/blockSize - square_of_mean);
 
 }
