+++ src/StatisticsFunctions/plp_var_q8.c
@@ -73,7 +73,7 @@ void plp_var_q8(
                          const int8_t * __restrict__ pSrc,
                          uint32_t blockSize,
                          uint32_t deciPoint,
-                         int32_t * __restrict__ pRes){
+                         int8_t * __restrict__ pRes){
   
   if (rt_cluster_id() == ARCHI_FC_CID){
     plp_var_q8s_rv32im(pSrc, blockSize, deciPoint, pRes);
